Added support for Maker badge
This is the last update for a while now. Fixed everything related to Maker badge. Voltage is measured as it should and everything just works. 
The only thing is, I haven't had a chance to test the functionality of the code for the D version, but it should work. 
Also, many thanks to everyone who helped me.